% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{calculate_time_difference}
\alias{calculate_time_difference}
\title{Calculate the time difference between two time points}
\usage{
calculate_time_difference(start_time, end_time)
}
\arguments{
\item{start_time}{The start time point.}

\item{end_time}{The end time point.}
}
\value{
A character string representing the elapsed time in minutes and seconds.
}
\description{
This function calculates the time difference between two given time points and returns
the result in minutes and seconds.
}
