% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_prepare.R
\name{scaleData_S2}
\alias{scaleData_S2}
\title{Scales Sentinel-2 imagery data}
\usage{
scaleData_S2(img)
}
\arguments{
\item{img}{An Earth Engine image object representing Sentinel-2 imagery.}
}
\value{
An Earth Engine image object with scaled pixel values.
}
\description{
This function scales the pixel values of Sentinel-2 imagery by multiplying them 
with a scaling factor of 0.0001. It is commonly used to convert the data from integer 
format to floating-point format.
}
\details{
The function performs the following step:
\enumerate{
  \item Multiplies the pixel values of the input image by a scaling factor of 0.0001 using 
  the \code{multiply} function.
}
}
