% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{spt_ba_data_url}
\alias{spt_ba_data_url}
\title{Burned Area Dataset URL}
\usage{
spt_ba_data_url(baDataset)
}
\arguments{
\item{baDataset}{The burned area dataset code (eg., EFFIS, ICNF).}
}
\value{
The URL of the specified burned area dataset.
}
\description{
This function retrieves the URL of the specified burned area dataset.
}
\examples{
url <- spt_ba_data_url("EFFIS")

}
