% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{spt_ba_dataset_code}
\alias{spt_ba_dataset_code}
\title{Burned Area Dataset Code}
\usage{
spt_ba_dataset_code(baDataset)
}
\arguments{
\item{baDataset}{The burnt area dataset name.}
}
\value{
Character. The dataset code corresponding to the given 
burned area dataset.
}
\description{
This function retrieves the code for a given burned area dataset.
}
\examples{
code <- spt_ba_dataset_code("ICNF")

}
