% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{spt_backup_file}
\alias{spt_backup_file}
\title{Create a Backup of a File}
\usage{
spt_backup_file(targetFile)
}
\arguments{
\item{targetFile}{A character string representing the file to be backed up.}
}
\description{
This function creates a backup of a specified file by appending the suffix "_bkp" to the file name.
}
\examples{
spt_backup_file("data.csv")
}
