% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_prepare.R
\name{spt_bitwise_extract}
\alias{spt_bitwise_extract}
\title{Extracts bits from a numeric value using bitwise operations}
\usage{
spt_bitwise_extract(value, fromBit, toBit = NULL)
}
\arguments{
\item{value}{A numeric value from which bits are to be extracted.}

\item{fromBit}{The starting bit position from which extraction should begin.}

\item{toBit}{(optional) The ending bit position at which extraction should stop.
If not provided, extraction will be performed for a single bit at
the \code{fromBit} position.}
}
\value{
A numeric value representing the extracted bits.
}
\description{
This function performs bitwise operations to extract a range of bits from a numeric value.
It can extract a single bit or a range of bits specified by the start and end positions.
}
\details{
The function performs bitwise right shift and bitwise AND operations to extract the
desired bits from the input value. If only \code{fromBit} is provided, a single bit is
extracted. If both \code{fromBit} and \code{toBit} are provided, a range of bits is extracted,
inclusive of both positions. The extracted bits are returned as a numeric value.
}
\examples{
# Extract a single bit
value <- 10  # Binary representation: 1010
extracted_bit <- spt_bitwise_extract(value, fromBit = 1)
extracted_bit  # Output: 0

# Extract a range of bits
value <- 10  # Binary representation: 1010
extracted_bits <- spt_bitwise_extract(value, fromBit = 2, toBit = 3)
extracted_bits  # Output: 1

}
