% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_indices.R
\name{spt_calc_bai}
\alias{spt_calc_bai}
\title{Calculate the Burned Area Index (BAI) for an image.}
\usage{
spt_calc_bai(img)
}
\arguments{
\item{img}{The input image for which to calculate the BAI.}
}
\value{
An image representing the calculated BAI with appropriate metadata.
}
\description{
This function calculates the Burned Area Index (BAI) for an image in Google Earth Engine (GEE).
}
\details{
The Burned Area Index (BAI) is a spectral index used to detect burned areas in remote sensing imagery.
It is calculated using the formula: 1 / ((0.1 - RED)^2 + (0.06 - NIR)^2),
where RED represents the red band values of the input image and NIR represents the near-infrared band values.
The BAI can be used to identify areas affected by fire and assess the severity of burn scars.
}
\examples{
# Assuming 'image' is a valid GEE image object
bai_image <- spt_calc_bai(image)

}
