% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_indices.R
\name{spt_calc_csi}
\alias{spt_calc_csi}
\title{Calculate Char Soil Index (CSI)}
\usage{
spt_calc_csi(img)
}
\arguments{
\item{img}{An input image containing NIR and SWIR2 bands.}
}
\value{
An image representing the CSI.
}
\description{
Calculates the Char Soil Index (CSI) from an input image using the NIR and SWIR2 
bands.
}
\details{
This function calculates the Char Soil Index (CSI) using the NIR and SWIR2 bands.
The formula for CSI is: NIR / SWIR2.

The input image should contain the NIR and SWIR2 bands as specified in the function.
The output image will have a band named 'CSI' and will inherit the properties of the 
input image.

# Smith et al. 2007
}
\examples{
# Load an image with NIR and SWIR2 bands
# img <- ee$Image('path/to/image')

# Calculate CSI
# csi <- spt_calc_csi(img)

}
