% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_indices.R
\name{spt_calc_dbsi}
\alias{spt_calc_dbsi}
\title{Calculate the Dry Bareness Index (DBSI) for an image.}
\usage{
spt_calc_dbsi(img)
}
\arguments{
\item{img}{The input image for which to calculate the DBSI.}
}
\value{
An image representing the calculated DBSI with appropriate metadata.
}
\description{
This function calculates the Dry Bareness Index (DBSI) for an image in Google Earth Engine (GEE).
}
\details{
The Dry Bareness Index (DBSI) is a spectral index used to quantify the dryness and bareness of an
area in remote sensing imagery.
It is calculated using the formula: ((SWIR1 - GREEN) / (SWIR1 + GREEN)) - ((NIR - RED) / (NIR + RED)),
where RED, NIR, SWIR1, and GREEN represent the respective band values of the input image.
The DBSI can be used to assess the degree of bareness and dryness of land surfaces, which is useful for
monitoring desertification, land degradation, and vegetation health.
}
\examples{
# Assuming 'image' is a valid GEE image object
dbsi_image <- spt_calc_dbsi(image)

}
