% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_indices.R
\name{spt_calc_evi}
\alias{spt_calc_evi}
\title{Calculate Enhanced Vegetation Index (EVI)}
\usage{
spt_calc_evi(img)
}
\arguments{
\item{img}{An input image containing NIR, Red, and Blue bands.}
}
\value{
An image representing the EVI.
}
\description{
Calculates the EVI from an input image using the NIR (Near Infrared), Red, and Blue bands.
}
\details{
This function calculates the Enhanced Vegetation Index (EVI) using the NIR, Red, and Blue bands.
The formula for EVI is: 2.5 ((NIR - Red) / (NIR + 6 Red - 7.5 Blue + 1)).

The input image should contain the NIR, Red, and Blue bands as specified in the function.
The output image will have a band named 'EVI' and will inherit the properties of the input image.
}
\examples{
# Load an image with NIR, Red, and Blue bands
# img <- ee$Image('path/to/image')

# Calculate EVI
# evi <- spt_calc_evi(img)

}
