% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_indices.R
\name{spt_calc_gemi}
\alias{spt_calc_gemi}
\title{Calculate the Global Environment Monitoring Index (GEMI) for an image.}
\usage{
spt_calc_gemi(img)
}
\arguments{
\item{img}{The input image for which to calculate the GEMI.}
}
\value{
An image representing the calculated GEMI with appropriate metadata.
}
\description{
This function calculates the Global Environment Monitoring Index (GEMI) for an image in
Google Earth Engine (GEE).
}
\details{
The Global Environment Monitoring Index (GEMI) is a spectral index used to estimate the greenness and
vegetation conditions in remote sensing.

It is calculated using the formula:
((2.0((NIR  2.0)-(RED  2.0)) + 1.5NIR + 0.5RED) /
(NIR + RED + 0.5))(1.0 - 0.25((2.0  ((NIR  2.0) - (RED  2)) + 1.5  NIR + 0.5  RED) /
(NIR + RED + 0.5)))-((RED - 0.125)/(1 - RED)),

where NIR and RED represent the respective near-infrared and red band values of the input image.

The GEMI can provide information about vegetation health, chlorophyll content, and overall environmental
monitoring.
}
\examples{
# Assuming 'image' is a valid GEE image object
gemi_image <- spt_calc_gemi(image)

}
