% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_indices.R
\name{spt_calc_mirbi}
\alias{spt_calc_mirbi}
\title{Calculate Mid-Infrared Burn Index (MIRBI)}
\usage{
spt_calc_mirbi(img)
}
\arguments{
\item{img}{An input image containing SWIR1 and SWIR2 bands.}
}
\value{
An image representing the MIRBI.
}
\description{
Calculates the Mid-Infrared Burn Index (MIRBI) from an input image using the SWIR1 and SWIR2 bands.
}
\details{
This function calculates the Mid-Infrared Burn Index (MIRBI) using the SWIR1 and SWIR2 bands.
The formula for MIRBI is: -1 * (10 * SWIR2 - 9.8 * SWIR1 + 2).

The input image should contain the SWIR1 and SWIR2 bands as specified in the function.
The output image will have a band named 'MIRBI' and will inherit the properties of the input image.

Trigg, S., & Flasse, S. (2001). An evaluation of different
bi-spectral spaces for discriminating burned shrub-savannah. International Journal
of RemoteSensing, 22(13), 2641-2647. https://doi.org/10.1080/01431160110053185

NOTE: THIS INDEX WAS ADJUSTED TO NEGATIVE FORM TO GIVE POSITIVE DELTAS HIGH SEVERITY
}
\examples{
# Load an image with SWIR1 and SWIR2 bands
# img <- ee$Image('path/to/image')

# Calculate MIRBI
# mirbi <- spt_calc_mirbi(img)

}
