% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_indices.R
\name{spt_calc_msavi}
\alias{spt_calc_msavi}
\title{Calculate the Modified Soil-Adjusted Vegetation Index (MSAVI) by Rogan and
Yool (2001) for an image.}
\usage{
spt_calc_msavi(img)
}
\arguments{
\item{img}{The input image for which to calculate the MSAVI.}
}
\value{
An image representing the calculated MSAVI with appropriate metadata.
}
\description{
This function calculates the Modified Soil-Adjusted Vegetation Index (MSAVI)
by Rogan and Yool (2001) for an image in Google Earth Engine (GEE).
}
\details{
The Modified Soil-Adjusted Vegetation Index (MSAVI) by Rogan and Yool (2001)
is a modification of the Soil-Adjusted Vegetation Index (SAVI) that improves the
index's performance in areas with low vegetation cover.

It is calculated using the formula:

(2NIR + 1 - sqrt(pow((2NIR + 1), 2) - 8 * (NIR - RED))) / 2,

where NIR and RED represent the respective near-infrared and red band values of
the input image.
The MSAVI is particularly useful for estimating vegetation greenness in sparse
vegetation areas.
}
\examples{
# Assuming 'image' is a valid GEE image object
msavi_image <- spt_calc_msavi(image)

}
