% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_indices.R
\name{spt_calc_nbr}
\alias{spt_calc_nbr}
\title{Calculate Normalized Burn Ratio (NBR)}
\usage{
spt_calc_nbr(img)
}
\arguments{
\item{img}{An input image containing NIR and SWIR2 bands.}
}
\value{
An image representing the NBR.
}
\description{
Calculates the NBR from an input image using the NIR (Near Infrared) and SWIR2 (Shortwave Infrared 2) bands.
}
\details{
This function calculates the NBR (Normalized Burn Ratio) by taking the difference
between the NIR and SWIR2 bands and normalizing it. The NBR is commonly used to assess
burned area severity and vegetation recovery after wildfires.

The input image should contain the NIR and SWIR2 bands as specified in the function.
The output image will have a band named 'NBR' and will inherit the properties of the input image.
}
\examples{
# Load an image with NIR and SWIR2 bands
# img <- ee$Image('path/to/image')

# Calculate NBR
# nbr <- spt_calc_nbr(img)

}
