% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_indices.R
\name{spt_calc_nbr2}
\alias{spt_calc_nbr2}
\title{Calculate the Normalized Burn Ratio 2 (NBR2) for an image}
\usage{
spt_calc_nbr2(img)
}
\arguments{
\item{img}{The input image for which to calculate the NBR2.}
}
\value{
An image representing the calculated NBR2 with appropriate metadata.
}
\description{
This function calculates the Normalized Burn Ratio 2 (NBR2) for an image in
Google Earth Engine (GEE).
}
\details{
The Normalized Burn Ratio 2 (NBR2) is a spectral index commonly used in remote sensing
to assess burned areas and monitor post-fire effects.
It is calculated using the formula: (SWIR1 - SWIR2) / (SWIR1 + SWIR2),
where SWIR1 and SWIR2 represent the respective short-wave infrared band values of
the input image.
The NBR2 can help identify and quantify the severity and extent of burn scars, vegetation
recovery, and changes in land cover after a fire event.
}
\examples{
# Assuming 'image' is a valid GEE image object
nbr2_image <- spt_calc_nbr2(image)

}
