% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_indices.R
\name{spt_calc_nbrp}
\alias{spt_calc_nbrp}
\title{Adapted NBR+ (Normalized Burn Ratio Plus) that are not Sentinel-2}
\usage{
spt_calc_nbrp(img)
}
\arguments{
\item{img}{An input image containing SWIR2, NIR, Green, and Blue bands.}
}
\value{
An image representing the NBR+.
}
\description{
Calculates the NBR+ (Normalized Burn Ratio Plus) from an input image using the SWIR2, NIR, Green,
and Blue bands. This function modifies the calculation to use the NIR band instead of the
Red edge-4 band used in the original index formulation.
}
\details{
This function calculates the NBR+ (Normalized Burn Ratio Plus) using the SWIR2, NIR, Green, and
Blue bands.
The formula for NBR+ is: -1 * ((SWIR2 - NIR - Green - Blue) / (SWIR2 + NIR + Green + Blue)).

The input image should contain the SWIR2, NIR, Green, and Blue bands as specified in the function.
The output image will have a band named 'NBRP' and will inherit the properties of the input image.

Please note that the NBR+ index was adjusted to a negative form to give positive deltas high severity.
The adjustment was made according to the paper by Alcaras (2022).
}
\examples{
# Load an image with SWIR2, NIR, Green, and Blue bands
# img <- ee$Image('path/to/image')

# Calculate NBR+
# nbrp <- spt_calc_nbrp(img)

}
\references{
Alcaras, M. 2022. "Satellite-Based Burn Severity Mapping: Implementation of the Normalized Burn
Ratio Plus (NBR+) for Sentinel-2 and Landsat-8."
Remote Sensing, 14(7), 1727. DOI: 10.3390/rs14071727.
}
