% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_indices.R
\name{spt_calc_nbrp_s2}
\alias{spt_calc_nbrp_s2}
\title{Calculate NBR+ (Normalized Burn Ratio Plus) for Sentinel-2}
\usage{
spt_calc_nbrp_s2(img)
}
\arguments{
\item{img}{An input image containing SWIR2, RE4, Green, and Blue bands.}
}
\value{
An image representing the NBR+ for Sentinel-2.
}
\description{
Calculates the NBR+ (Normalized Burn Ratio Plus) for Sentinel-2 images using the SWIR2, 
RE4, Green, and Blue bands.
}
\details{
This function calculates the NBR+ (Normalized Burn Ratio Plus) for Sentinel-2 images using 
the SWIR2, RE4, Green, and Blue bands.
The formula for NBR+ is: -1 * ((SWIR2 - RE4 - Green - Blue) / (SWIR2 + RE4 + Green + Blue)).

The input image should contain the SWIR2, RE4, Green, and Blue bands as specified in the function.
The output image will have a band named 'NBRP' and will inherit the properties of the input image.

Note: Please note that the NBR+ index was adjusted to a negative form to give positive deltas high severity.
The adjustment was made according to the specified note.
}
\examples{
# Load an image with SWIR2, RE4, Green, and Blue bands
# img <- ee$Image('path/to/image')

# Calculate NBR+ for Sentinel-2
# nbrp <- spt_calc_nbrp_s2(img)

}
\references{
Alcaras, M. 2022. "Satellite-Based Burn Severity Mapping: Implementation of the Normalized Burn 
Ratio Plus (NBR+) for Sentinel-2 and Landsat-8." 
Remote Sensing, 14(7), 1727. DOI: 10.3390/rs14071727.
}
