% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_indices.R
\name{spt_calc_nbrswir}
\alias{spt_calc_nbrswir}
\title{Calculate a novel Normalized Burn Ratio (NBR_SWIR)}
\usage{
spt_calc_nbrswir(img)
}
\arguments{
\item{img}{An input image containing SWIR1 and SWIR2 bands.}
}
\value{
An image representing the NBR_SWIR index.
}
\description{
Calculates the Normalized Burn Ratio (NBR_SWIR) from an input image using the SWIR1 and SWIR2 bands.
}
\details{
This function calculates the Normalized Burn Ratio (NBR_SWIR) using the SWIR1 and SWIR2 bands.
The formula for NBR_SWIR is: - ((SWIR2 - SWIR1 - 0.02) / (SWIR2 + SWIR1 + 0.1)).

The input image should contain the SWIR1 and SWIR2 bands as specified in the function.
The output image will have a band named 'NBRSWIR' and will inherit the properties of the input image.

Based on:
Sicong Liu , Yongjie Zheng , Michele Dalponte & Xiaohua Tong (2020) A
novel fire index-based burned area change detection approach using Landsat-8 OLI data,
European Journal of Remote Sensing, 53:1, 104-112, DOI: 10.1080/22797254.2020.1738900

NOTE: THIS INDEX WAS ADJUSTED TO NEGATIVE FORM TO GIVE POSITIVE DELTAS HIGH SEVERITY
}
\examples{
# Load an image with SWIR1 and SWIR2 bands
# img <- ee$Image('path/to/image')

# Calculate NBR_SWIR
# nbr_swir <- spt_calc_nbrswir(img)

}
