% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_indices.R
\name{spt_calc_ndvi}
\alias{spt_calc_ndvi}
\title{Calculate Normalized Difference Vegetation Index (NDVI)}
\usage{
spt_calc_ndvi(img)
}
\arguments{
\item{img}{An input image containing NIR and Red bands.}
}
\value{
An image representing the NDVI.
}
\description{
Calculates the NDVI from an input image using the NIR (Near Infrared) and Red bands.
}
\details{
This function calculates the NDVI (Normalized Difference Vegetation Index) by taking the difference
between the NIR and Red bands and normalizing it. The NDVI is a common index used to assess vegetation
health and density.

The input image should contain the NIR and Red bands as specified in the function.
The output image will have a band named 'NDVI' and will inherit the properties of the input image.
}
\examples{
# Load an image with NIR and Red bands
# img <- ee$Image('path/to/image')

# Calculate NDVI
# ndvi <- spt_calc_ndvi(img)

}
