% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_indices.R
\name{spt_calc_ndwi}
\alias{spt_calc_ndwi}
\title{Calculate the Normalized Difference Water Index (NDWI) for an image.}
\usage{
spt_calc_ndwi(img)
}
\arguments{
\item{img}{The input image for which to calculate the NDWI.}
}
\value{
An image representing the calculated NDWI with appropriate metadata.
}
\description{
This function calculates the Normalized Difference Water Index (NDWI) for an
image in Google Earth Engine (GEE).
}
\details{
The Normalized Difference Water Index (NDWI) is a derived index used to estimate the
leaf water content at the canopy level. It is particularly useful for mapping water
bodies and assessing vegetation water stress. The NDWI is calculated using the formula:

(NIR - SWIR1) / (NIR + SWIR1)

where NIR and SWIR1 represent the respective near-infrared and shortwave infrared
band values of the input image.
Higher NDWI values indicate the presence of more water content, while lower values
indicate less water content.

Reference: [NDWI Factsheet](https://edo.jrc.ec.europa.eu/documents/factsheets/factsheet_ndwi.pdf)
}
\examples{
# Assuming 'image' is a valid GEE image object
ndwi_image <- spt_calc_ndwi(image)

}
