% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_indices.R
\name{spt_calc_savi}
\alias{spt_calc_savi}
\title{Calculate the Soil-Adjusted Vegetation Index (SAVI) for an image.}
\usage{
spt_calc_savi(img)
}
\arguments{
\item{img}{The input image for which to calculate the SAVI.}
}
\value{
An image representing the calculated SAVI with appropriate metadata.
}
\description{
This function calculates the Soil-Adjusted Vegetation Index (SAVI) for an image in
Google Earth Engine (GEE).
}
\details{
The Soil-Adjusted Vegetation Index (SAVI) is a vegetation index that adjusts the
standard NDVI (Normalized Difference Vegetation Index) by taking into account the
soil background reflectance. It is calculated using the formula:
(1 + L) * (NIR - RED) / (NIR + RED + L),

where NIR and RED represent the respective near-infrared and red band values of the
input image, and L is a soil adjustment factor (typically set to 0.5).

The SAVI is useful for vegetation analysis in areas with varying soil backgrounds
and can help minimize the influence of soil reflectance.
}
\examples{
# Assuming 'image' is a valid GEE image object
savi_image <- spt_calc_savi(image)

}
