% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_indices.R
\name{spt_calc_tctb_l5}
\alias{spt_calc_tctb_l5}
\title{Calculate TCTB (Tasseled Cap Transformation Brightness) for Landsat-5}
\usage{
spt_calc_tctb_l5(img)
}
\arguments{
\item{img}{An input image containing the Blue, Green, Red, NIR, SWIR1, and SWIR2 bands.}
}
\value{
An image containing the TCTB values.
}
\description{
Calculates the TCTB (Tasseled Cap Transformation Brightness) for Landsat-5 images using
the Blue, Green, Red, NIR, SWIR1, and SWIR2 bands.
}
\details{
The TCTB is calculated using the formula described in the references.
}
