% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_indices.R
\name{spt_calc_tctg_s2}
\alias{spt_calc_tctg_s2}
\title{Calculate TCTG (Tasseled Cap Transformation Greenness) for Sentinel-2}
\usage{
spt_calc_tctg_s2(img)
}
\arguments{
\item{img}{An input image containing the Blue, Green, Red, NIR, SWIR1, and SWIR2 bands.}
}
\value{
An image containing the TCTG values.
}
\description{
Calculates the TCTG (Tasseled Cap Transformation Greenness) for Sentinel-2 images using the 
Blue, Green, Red, NIR, SWIR1, and SWIR2 bands.
}
\details{
The TCTG is calculated using the following formula:
\deqn{TCTG = (BLUE \times -0.3599) + (GREEN \times -0.3533) + (RED \times -0.4734) + 
(NIR \times 0.6633) + (SWIR1 \times 0.0087) + (SWIR2 \times -0.2856)}
}
\examples{
# Load a Sentinel-2 image
image <- ee$Image('COPERNICUS/S2_SR/20210901T184201_20210901T184205_T18TVJ')

# Calculate TCTG
tctg <- spt_calc_tctg_s2(image)

}
