% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_indices.R
\name{spt_calc_tctw_l7}
\alias{spt_calc_tctw_l7}
\title{Calculate TCTW (Tasseled Cap Transformation Wetness) for Landsat-7}
\usage{
spt_calc_tctw_l7(img)
}
\arguments{
\item{img}{An input image containing the Blue, Green, Red, NIR, SWIR1, and SWIR2 bands.}
}
\value{
An image containing the TCTW values.
}
\description{
Calculates the TCTW (Tasseled Cap Transformation Wetness) for Landsat-7 images using the
Blue, Green, Red, NIR, SWIR1, and SWIR2 bands.
}
\details{
The TCTW is calculated using a specific formula.
}
