% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_indices.R
\name{spt_calc_tctw_mod09a1}
\alias{spt_calc_tctw_mod09a1}
\title{Calculate TCTW (Tasseled Cap Transformation Wetness) for MOD09A1}
\usage{
spt_calc_tctw_mod09a1(img)
}
\arguments{
\item{img}{An input image containing the Blue, Green, Red, NIR, NIR2, SWIR1, and 
SWIR2 bands.}
}
\value{
An image containing the TCTW values.
}
\description{
Calculates the TCTW (Tasseled Cap Transformation Wetness) for MOD09A1 images 
using the Blue, Green, Red, NIR, NIR2, SWIR1, and SWIR2 bands.
}
\details{
The TCTW is calculated using a specific formula.
}
