% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_generator.R
\name{spt_check_values}
\alias{spt_check_values}
\title{Check Validity of Parameter Values}
\usage{
spt_check_values(value, what, verbose = TRUE)
}
\arguments{
\item{value}{The value(s) to be checked.}

\item{what}{A character string specifying the parameter name or type.}

\item{verbose}{A logical value indicating whether to display a message when a parameter cannot be checked. Default is TRUE.}
}
\value{
A logical value indicating whether the value(s) pass the validity check.
}
\description{
This function checks the validity of parameter values based on predefined criteria or value ranges.
}
