% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_calc.R
\name{spt_cloud_mask_fun}
\alias{spt_cloud_mask_fun}
\title{Get a specific Cloud Mask Function by satellite name}
\usage{
spt_cloud_mask_fun(satCode, modisProduct = NULL)
}
\arguments{
\item{satCode}{The satellite code (character).}

\item{modisProduct}{The MODIS product (character). Default is NULL.}
}
\value{
The cloud mask function.
}
\description{
This function selects and returns the corresponding cloud mask function based on the specified parameters.
}
