% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{spt_create_file}
\alias{spt_create_file}
\title{Create File with Current Timestamp}
\usage{
spt_create_file(filename)
}
\arguments{
\item{filename}{A character string representing the name of the file to be created.}
}
\description{
This function creates a new file with the specified filename and writes the current timestamp to it.
If a file with the same name already exists, it will be removed before creating the new file.
}
\examples{
spt_create_file("output.txt")
}
