% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{spt_current_date_time}
\alias{spt_current_date_time}
\title{Get the Current Date and Time}
\usage{
spt_current_date_time()
}
\value{
A character string representing the current date and time in "YYYY-MM-DD HH:MM:SS" format.
}
\description{
This function returns the current date and time in the format "YYYY-MM-DD HH:MM:SS".
}
\examples{
spt_current_date_time()
}
