% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{spt_current_year}
\alias{spt_current_year}
\title{Get the Current Year}
\usage{
spt_current_year()
}
\value{
A character string representing the current year.
}
\description{
This function returns the current year as a character string.
}
\examples{
spt_current_year()
}
