% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{spt_df_to_md}
\alias{spt_df_to_md}
\title{Convert Data Frame to Markdown Table}
\usage{
spt_df_to_md(df)
}
\arguments{
\item{df}{A data frame to be converted to Markdown.}
}
\value{
A knitr_kable object representing the data frame in Markdown table format.
}
\description{
This function converts a data frame into a Markdown table format.
}
\examples{
spt_df_to_md(iris)
}
