% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{spt_download_unzip}
\alias{spt_download_unzip}
\title{Download and Unzip File}
\usage{
spt_download_unzip(url, dest_file, extdir)
}
\arguments{
\item{url}{A character string representing the URL of the file to be downloaded.}

\item{dest_file}{A character string representing the destination file path where the downloaded file will be saved.}

\item{extdir}{A character string representing the path to extract the contents of the ZIP file. Default is the current working directory.}
}
\description{
This function downloads a file from the specified URL and optionally unzips it if it is a ZIP file.
}
