% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_data.R
\name{spt_etm_to_oli}
\alias{spt_etm_to_oli}
\title{Radiometric corrections between Landsat TM/ETM and OLI sensors}
\usage{
spt_etm_to_oli(img)
}
\arguments{
\item{img}{An Earth Engine image object representing Landsat TM/ETM+ imagery.}
}
\value{
An Earth Engine image object representing Landsat OLI reflectance.
}
\description{
Converts/merges Landsat Thematic Mapper (TM) and Enhanced Thematic Mapper Plus (ETM+)
imagery to Operational Land Imager (OLI) reflectance using the provided coefficients.
}
\examples{
# Convert Landsat ETM+ image to OLI reflectance
etm_image <- ee$Image('LANDSAT/LE07/C01/T1_SR/LE07_044034_20181028')
oli_image <- spt_etm_to_oli(etm_image)

}
