% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{spt_export_meta_to_json}
\alias{spt_export_meta_to_json}
\title{Export Metadata to JSON}
\usage{
spt_export_meta_to_json(x, outFilePath)
}
\arguments{
\item{x}{The data frame to be exported.}

\item{outFilePath}{The file path for the JSON output file.}
}
\description{
This function exports a data frame to a prettified JSON file.
}
\examples{
spt_export_meta_to_json(myDataFrame, "path/to/output.json")

}
