% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{spt_export_to_md}
\alias{spt_export_to_md}
\title{Export Data Frame to Markdown File}
\usage{
spt_export_to_md(df, filename)
}
\arguments{
\item{df}{A data frame to be exported to Markdown.}

\item{filename}{A character string representing the name of the Markdown file.}
}
\description{
This function exports a data frame to a Markdown file.
}
\examples{
spt_export_to_md(iris, "output.md")
}
