% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{spt_file_dates}
\alias{spt_file_dates}
\title{Check File Creation and Modification Dates}
\usage{
spt_file_dates(filename, tdelta = 60)
}
\arguments{
\item{filename}{A character string representing the name of the file to check.}

\item{tdelta}{The time interval in minutes. If the file's creation or modification date is within this interval
from the current time, TRUE is returned. Default is 60 minutes.}
}
\value{
A logical value indicating whether the file's creation or modification date falls within the specified time interval.
}
\description{
This function checks the creation and modification dates of a file and returns TRUE if either of the dates
falls within a specified time interval from the current time.
}
\examples{
spt_file_dates("data.txt", tdelta = 60)
spt_file_dates("data.txt")
}
