% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{spt_fill_meta_template}
\alias{spt_fill_meta_template}
\title{Fill Metadata Template}
\usage{
spt_fill_meta_template(metaTemplate, metaList)
}
\arguments{
\item{metaTemplate}{The metadata template data frame.}

\item{metaList}{The metadata list containing parameter-value pairs.}
}
\value{
The filled metadata template with updated values.
}
\description{
This function fills in a metadata template with values from a metadata list.
}
\examples{
filledTemplate <- spt_fill_meta_template(metaTemplate, metaList)

}
