% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_utils.R
\name{spt_gee_task_status}
\alias{spt_gee_task_status}
\title{Retrieve the status of a Google Earth Engine (GEE) task and convert it to an R list object}
\usage{
spt_gee_task_status(geeTask)
}
\arguments{
\item{geeTask}{The GEE task object.}
}
\value{
A list containing the task status and additional parameters for the download function.
}
\description{
This function retrieves the status of a GEE task and converts it to an R list object.
}
\details{
The function checks the status of the GEE task using \code{ee$batch$Task$status(geeTask)} and converts
the result from Python to an R list object using the \code{ee_utils_py_to_r} function. It then adds
extra parameters to the output list, such as the file export options, filename prefix, and file format,
which are needed for the download function but cannot be stored directly from the GEE task object.

If the task has finished and the \code{"destination_uris"} field exists in the output list, the function
retrieves the basename of the URI for the Google Drive file and stores it as \code{basename_uri}.
}
\examples{
# Assuming 'geeTask' is a valid GEE task object
spt_gee_task_status(geeTask)

}
\seealso{
\code{\link{ee_utils_py_to_r}}, \code{\link{basename}}
}
