% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_generator.R
\name{spt_generate_tasks}
\alias{spt_generate_tasks}
\title{Generate Tasks}
\usage{
spt_generate_tasks(
  taskTable = NULL,
  satCode,
  procLevel,
  modisProduct,
  baseIndex,
  severityIndicator,
  burntAreaDataset,
  referenceYear,
  preFireRef,
  preFireType,
  postFireRef,
  minFireSize
)
}
\arguments{
\item{taskTable}{An optional existing tasks table to append the generated tasks to.}

\item{satCode}{The satellite code.}

\item{procLevel}{The processing level.}

\item{modisProduct}{The MODIS product.}

\item{baseIndex}{The base index.}

\item{severityIndicator}{The severity indicator.}

\item{burntAreaDataset}{The burnt area dataset.}

\item{referenceYear}{The reference year.}

\item{preFireRef}{The pre-fire reference period in months.}

\item{preFireType}{The type of pre-fire reference period.}

\item{postFireRef}{The post-fire reference period in months.}

\item{minFireSize}{The minimum fire size.}
}
\value{
A data frame representing the tasks table with appended rows for the generated tasks.
}
\description{
This function generates a table of tasks based on the provided analysis parameters.
}
