% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_data.R
\name{spt_get_ba_dataset}
\alias{spt_get_ba_dataset}
\title{Get and pre-process the Burned Area Dataset}
\usage{
spt_get_ba_dataset(
  baDataset,
  baGEEasset,
  referenceYear,
  minFireSizeHa = 5,
  dateField,
  yearField,
  areaField
)
}
\arguments{
\item{baDataset}{The dataset name ("ICNF" or "EFFIS").}

\item{baGEEasset}{The GEE asset representing the burned area dataset.}

\item{referenceYear}{The reference year for retrieving BA data.}

\item{minFireSizeHa}{The minimum fire size in hectares to filter the BA data (default: 5).}

\item{dateField}{The field name representing the date in the BA dataset.}

\item{yearField}{The field name representing the year in the BA dataset.}

\item{areaField}{The field name representing the area in hectares in the BA dataset.}
}
\value{
The burned area data as an ee.FeatureCollection object.
}
\description{
This function retrieves burned area (BA) data from the specified dataset and
Google Earth Engine (GEE) asset. Performs a set of filters needed to remove NA's
and the minimum area size.
}
\examples{
baData <- spt_get_ba_dataset("ICNF", "users/username/ICNF_BA", 2022, 10, "date", "year", "area")

}
