% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_prepare.R
\name{spt_lt7_interpolate}
\alias{spt_lt7_interpolate}
\title{Interpolates missing values in Landsat 7 imagery}
\usage{
spt_lt7_interpolate(img)
}
\arguments{
\item{img}{An Earth Engine image object representing Landsat 7 imagery.}
}
\value{
An Earth Engine image object with missing values interpolated using a focal mean filter.
}
\description{
This function applies a focal mean filter to the Landsat 7 imagery in order to 
interpolate missing values due to SLC-off issues. It calculates the average of 
neighboring pixels within a square kernel of size 8x8 and replaces the missing 
alues with the interpolated values.
}
\details{
The function uses the \code{focal_mean} function to calculate the average of neighboring 
pixels within a square kernel of size 8x8. This kernel is applied to the input image, and 
the resulting values are used to replace the missing values in the image. The interpolation 
helps fill in the gaps caused by missing or corrupted data in the Landsat 7 imagery caused by 
SLC-off.
}
