% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_generator.R
\name{spt_make_tasks_table}
\alias{spt_make_tasks_table}
\title{Create Tasks Table}
\usage{
spt_make_tasks_table(n)
}
\arguments{
\item{n}{The number of rows to create in the tasks table.}
}
\value{
A data frame representing the tasks table with empty columns.
}
\description{
This function creates an empty tasks table with predefined columns to store task-related information.
}
