% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_prepare.R
\name{spt_mask_clouds_mod09a1}
\alias{spt_mask_clouds_mod09a1}
\title{Masks clouds, cloud shadows, and cirrus in MOD09A1 imagery}
\usage{
spt_mask_clouds_mod09a1(image)
}
\arguments{
\item{image}{An ee$Image object representing the MOD09A1 imagery.}
}
\value{
An ee$Image object with clouds, cloud shadows, and cirrus masked out.
}
\description{
This function applies cloud, cloud shadow, and cirrus masks to MOD09A1 imagery
based on the StateQA band values. It extracts specific bits from the StateQA band
using the \code{spt_bitwise_extract} function and applies masks based on the extracted values.
}
\details{
The function applies masks to the input image using the StateQA band values. It extracts
specific bits from the StateQA band using the \code{spt_bitwise_extract} function and creates
masks based on the extracted values. Clouds, cloud shadows, and cirrus are masked out by setting
the corresponding pixels to \code{NA} (Not Available) in the output image.
}
