% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{spt_pad_number}
\alias{spt_pad_number}
\title{Pad a Number with Leading Zeros}
\usage{
spt_pad_number(x, width = 3)
}
\arguments{
\item{x}{A numeric or character value to be padded.}

\item{width}{An integer specifying the desired width of the padded number. Default is 3.}
}
\value{
A character string representing the padded number.
}
\description{
This function pads a number with leading zeros to a specified width.
}
\examples{
spt_pad_number(7, width = 5)
spt_pad_number(123, width = 8)
spt_pad_number("42", width = 4)
}
