% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{spt_periods_to_today}
\alias{spt_periods_to_today}
\title{Calculate the Number of Periods from a Date to Today}
\usage{
spt_periods_to_today(date, ndays)
}
\arguments{
\item{date}{A character or Date object representing the starting date.}

\item{ndays}{An integer specifying the number of days per period.}
}
\value{
An integer indicating the number of periods from the starting date to today.
}
\description{
This function calculates the number of periods from a given date to today, based on a specified number of days per period.
}
\examples{
spt_periods_to_today("2022-01-01", 7)
spt_periods_to_today(as.Date("2021-12-15"), 30)
}
