% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_management.R
\name{spt_post_window_days.task}
\alias{spt_post_window_days.task}
\title{Post-Fire Window Days for Task}
\usage{
\method{spt_post_window_days}{task}(x)
}
\arguments{
\item{x}{A task object containing information about the pre-fire and post-fire references.}
}
\value{
A numeric vector with two elements: the number of days in the post-fire window 
from pre-fire reference to post-fire reference, and the number of days in the post-fire 
reference.
}
\description{
This function calculates the number of days in the post-fire window for a given task, 
based on the pre-fire and post-fire references.
}
\examples{
task <- spt_get_task(taskTable = myTaskTable, taskID = "12345")
spt_post_window_days.task(task)

}
