% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{spt_proc_levels}
\alias{spt_proc_levels}
\title{Processing Levels description}
\usage{
spt_proc_levels(procLevel)
}
\arguments{
\item{procLevel}{The processing level of the satellite data.}
}
\value{
A description of the processing level.
}
\description{
This function provides a description of the processing levels for satellite data.
}
\examples{
levelDesc <- spt_proc_levels("L1C")

}
