% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_calc.R
\name{spt_process_gee_task}
\alias{spt_process_gee_task}
\title{Process GEE Task}
\usage{
spt_process_gee_task(
  task,
  boundBox,
  coordRefSys,
  baGEEasset,
  dateField,
  yearField,
  areaField,
  outFolder
)
}
\arguments{
\item{task}{An object representing the GEE task.}

\item{boundBox}{A bounding box specifying the spatial extent of the analysis.}

\item{coordRefSys}{The coordinate reference system (CRS) code for the analysis.}

\item{baGEEasset}{The GEE asset containing the burned area data.}

\item{dateField}{The name of the field in the burned area data representing the date.}

\item{yearField}{The name of the field in the burned area data representing the year.}

\item{areaField}{The name of the field in the burned area data representing the area.}

\item{outFolder}{The output folder where the processed data will be saved.}
}
\value{
None.
}
\description{
This function processes a GEE task by performing various calculations and operations
on satellite imagery and burned area data.
}
