% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{spt_product_name}
\alias{spt_product_name}
\title{Generate the Product Name}
\usage{
spt_product_name(
  ProjectAccronym,
  SeverityIndicator,
  BaseIndex,
  SatCode,
  BurntAreaDataset,
  ReferenceYear,
  RefPeriods,
  PrimaryCRScode,
  addCalcDate = FALSE,
  VersionNumber
)
}
\arguments{
\item{ProjectAccronym}{A character string representing the project acronym.}

\item{SeverityIndicator}{A character string representing the severity indicator.}

\item{BaseIndex}{A character string representing the base index.}

\item{SatCode}{A character string representing the satellite code.}

\item{BurntAreaDataset}{A character string representing the burnt area dataset.}

\item{ReferenceYear}{A character string representing the reference year.}

\item{RefPeriods}{A character string representing the reference periods.}

\item{PrimaryCRScode}{A character string representing the primary CRS code.}

\item{addCalcDate}{A logical value indicating whether to include the calculation date in the product name. Default is FALSE.}

\item{VersionNumber}{A character string representing the version number.}
}
\value{
A character string representing the generated product name.
}
\description{
This function generates a product name based on the provided parameters.
}
\examples{
spt_product_name("PROJ", "SEV", "BASE", "SAT", "BA", "2022", "PERIODS", "CRS", addCalcDate = TRUE, "V1.0")
spt_product_name("PROJ", "SEV", "BASE", "SAT", "BA", "2022", "PERIODS", "CRS", "V1.0")
}
