% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_proc.R
\name{spt_project_to_pttm06}
\alias{spt_project_to_pttm06}
\title{Project Raster to ETRS89/PT TM06 Coordinate Reference System}
\usage{
spt_project_to_pttm06(x, writeData = TRUE, outPath = NULL, ...)
}
\arguments{
\item{x}{A raster object or a character string representing the path to the raster file.}

\item{writeData}{A logical value indicating whether to write the projected raster to a 
new file. Default is TRUE.}

\item{outPath}{A character string representing the output file path for the projected 
raster. Ignored if writeData is FALSE.}

\item{...}{Additional arguments to be passed to the project function.}
}
\value{
If writeData is TRUE, the function returns a raster object representing the 
projected raster. Otherwise, it returns NULL.
}
\description{
This function projects a raster object or a raster file to the ETRS89/PT TM06 (EPSG:3763) 
coordinate reference system.
It can optionally write the projected raster to a new file with the updated CRS code in 
the filename.
}
