% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_proc.R
\name{spt_raster_zeros_to_na}
\alias{spt_raster_zeros_to_na}
\title{Convert Zeros to NA in Raster}
\usage{
spt_raster_zeros_to_na(rstPath, writeRast = TRUE, ...)
}
\arguments{
\item{rstPath}{A character string representing the path to the raster file.}

\item{writeRast}{A logical value indicating whether to write the modified raster 
to the same file. Default is TRUE.}

\item{...}{Additional arguments to be passed to the writeRaster function if 
writeRast is TRUE.}
}
\value{
If writeRast is TRUE, the function returns a raster object representing 
the modified raster. Otherwise, it returns NULL.
}
\description{
This function reads a raster file, converts all zero values to NA (missing values), 
and optionally writes the modified raster to the same file.
}
