% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_utils.R
\name{spt_read_gee_status_list}
\alias{spt_read_gee_status_list}
\title{Read a Google Earth Engine (GEE) status list from disk.}
\usage{
spt_read_gee_status_list(task, geeTaskPath)
}
\arguments{
\item{task}{The GEE task object.}

\item{geeTaskPath}{The path to the GEE task folder where the status list is stored.}
}
\value{
The GEE status list read from disk.
}
\description{
This function reads a GEE status list from disk in RDS format.
}
\details{
The function reads the GEE status list from disk using the \code{readRDS} function. It assumes
that the status list was previously saved in the specified \code{geeTaskPath} using the
\code{spt_save_gee_status_list} function.
}
