% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_generator.R
\name{spt_read_tasks_table}
\alias{spt_read_tasks_table}
\title{Read Tasks Table}
\usage{
spt_read_tasks_table(taskTablePath)
}
\arguments{
\item{taskTablePath}{The file path of the task table.}
}
\value{
A data frame representing the read task table.
}
\description{
This function reads a task table from a specified file path and returns the table as a data frame.
Uses a file lock.
}
