% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{spt_replace_crs_code}
\alias{spt_replace_crs_code}
\title{Replace the Coordinate Reference System (CRS Code) in a String}
\usage{
spt_replace_crs_code(x, from = "32629", to = "3763")
}
\arguments{
\item{x}{A character string in which the CRS code will be replaced.}

\item{from}{A character string representing the CRS code to be replaced. Default is "32629".}

\item{to}{A character string representing the new CRS code to replace the original CRS code. Default is "3763".}
}
\value{
A character string with the replaced CRS code.
}
\description{
This function replaces a specific CRS (Coordinate Reference System) code in a given string with a new CRS code.
The default CRS codes used are "32629" for the source CRS and "3763" for the target CRS.
}
\examples{
spt_replace_crs_code("data_32629.csv")
spt_replace_crs_code("data_32629.csv", from = "32629", to = "4326")
}
