% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_management.R
\name{spt_rm_uncompleted_tasks}
\alias{spt_rm_uncompleted_tasks}
\title{Remove Uncompleted Tasks}
\usage{
spt_rm_uncompleted_tasks(taskTable = NULL, taskTablePath)
}
\arguments{
\item{taskTable}{The task table to remove uncompleted tasks from (optional).
If not provided, it will be read from the specified \code{taskTablePath}.}

\item{taskTablePath}{The file path to the task table.}
}
\value{
\code{TRUE} if the removal is successful, \code{FALSE} otherwise.
}
\description{
This function removes uncompleted tasks from the task table.
}
\examples{
spt_rm_uncompleted_tasks(taskTable = myTaskTable, taskTablePath = "path/to/taskTable.csv")

}
