% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_management.R
\name{spt_sat_code}
\alias{spt_sat_code}
\title{Define the generic methods to access task parameters}
\usage{
spt_sat_code(x)
}
\arguments{
\item{x}{An object for which the satellite analysis operation is to be performed.}

\item{...}{Additional arguments to be passed to the dispatched method.}
}
\description{
This function is a dispatch method that performs various satellite analysis
operations based on the input object. The specific method to be executed is determined
by the class of the input object.
}
