% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_management.R
\name{spt_sat_code.task}
\alias{spt_sat_code.task}
\title{Define the get functions to access task parameters}
\usage{
\method{spt_sat_code}{task}(x)
}
\description{
Define the get functions to access task parameters
}
