% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{spt_sat_mission_fullname}
\alias{spt_sat_mission_fullname}
\title{Satellite Mission Full Name}
\usage{
spt_sat_mission_fullname(satCode)
}
\arguments{
\item{satCode}{The code of the satellite}
}
\value{
The full name of the satellite mission corresponding to the given code.
}
\description{
This function retrieves the full name of a satellite mission based on its code.
}
\examples{
fullname <- spt_sat_mission_fullname("S2MSI")

}
