% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_utils.R
\name{spt_save_gee_status_list}
\alias{spt_save_gee_status_list}
\title{Save a Google Earth Engine (GEE) status list to disk.}
\usage{
spt_save_gee_status_list(task, geeStatusList, geeTaskPath)
}
\arguments{
\item{task}{The GEE task object.}

\item{geeStatusList}{The GEE status list to be saved.}

\item{geeTaskPath}{The path to the GEE task folder where the files will be saved.}
}
\description{
This function saves a GEE status list to disk as RDS and CSV files.
}
\details{
The function saves the GEE status list to disk in two formats: RDS and CSV. It uses the \code{saveRDS} 
function to save the status list as an RDS file, and \code{write.csv} to save it as a CSV file. 
The files are named based on the task ID andtask UID to ensure uniqueness.
}
