% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_calc.R
\name{spt_scale_fun}
\alias{spt_scale_fun}
\title{Get a specific Scale Function based on the satellite code and processing level}
\usage{
spt_scale_fun(satCode, procLevel = NULL, modisProduct = NULL)
}
\arguments{
\item{satCode}{The satellite code (character).}

\item{procLevel}{The processing level (character). Default is NULL.}

\item{modisProduct}{The MODIS product (character). Default is NULL.}
}
\value{
The scaling function.
}
\description{
This function selects and returns the corresponding scaling function based on 
the specified parameters.
}
