% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_prepare.R
\name{spt_scale_lt_toa}
\alias{spt_scale_lt_toa}
\title{Scales Landsat top-of-atmosphere (TOA) imagery (placeholder function)}
\usage{
spt_scale_lt_toa(img)
}
\arguments{
\item{img}{An Earth Engine image object representing Landsat top-of-atmosphere (TOA) imagery.}
}
\value{
An Earth Engine image object with the TOA reflectance values
}
\description{
This function maintains the top-of-atmosphere (TOA) reflectance (placeholder function because values
area already scaled in this case)
}
\details{
The function simply multiplies the input image by a scaling factor of 1 using the \code{multiply} function.
This operation retains the original TOA values of the image.
}
