% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_prepare.R
\name{spt_scale_mod}
\alias{spt_scale_mod}
\title{Scales the values of MODIS images}
\usage{
spt_scale_mod(img)
}
\arguments{
\item{img}{An `ee$Image` object representing the MOD-format image.}
}
\value{
An `ee$Image` object with scaled pixel values.
}
\description{
This function scales the pixel values of an image in MOD format by multiplying
them with a scaling factor of 0.0001.
}
\details{
The function applies a scaling factor of 0.0001 to the pixel values of the input image.
This scaling is commonly used to convert MOD-format pixel values to their physical
measurement units.
}
