% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{spt_severity_indicator_form}
\alias{spt_severity_indicator_form}
\title{Get the Severity Indicator Formula}
\usage{
spt_severity_indicator_form(spi, si)
}
\arguments{
\item{spi}{The selected spectral index.}

\item{si}{The selected severity indicator.}
}
\value{
The formula for the severity indicator based on the selected
spectral index and severity indicator.
}
\description{
This function generates the formula for a severity indicator based on the
selected spectral index (spi) and severity indicator (si). Available options
are delta, Relative delta and Relativized Burn Ratio.
}
\examples{
formula <- spt_severity_indicator_form("NDVI", "DELTA")

}
