% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_management.R
\name{spt_spatial_resolution.task}
\alias{spt_spatial_resolution.task}
\title{Spatial Resolution for Task}
\usage{
\method{spt_spatial_resolution}{task}(x)
}
\arguments{
\item{x}{A task object containing information about the satellite code and the MODIS 
product (if applicable).}
}
\value{
The spatial resolution in meters.
}
\description{
This function determines the spatial resolution for a given task based on the satellite 
code and MODIS product (if applicable).
}
\examples{
task <- spt_get_task(taskTable = myTaskTable, taskID = "12345")
spt_spatial_resolution.task(task)

}
