% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{spt_spec_index_formula}
\alias{spt_spec_index_formula}
\title{Spectral Index Formula}
\usage{
spt_spec_index_formula(spiAcronym, satCode)
}
\arguments{
\item{spiAcronym}{The acronym of the spectral index.}

\item{satCode}{The satellite code.}
}
\value{
Character. The formula of the spectral index corresponding to the given 
acronym and satellite code.
}
\description{
This function retrieves the formula of a spectral index based on its 
acronym and satellite code.
}
\examples{
formula <- spt_spec_index_formula("TCTB", "MOD")

}
