% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{spt_spec_index_fullname}
\alias{spt_spec_index_fullname}
\title{Spectral Index Full Name}
\usage{
spt_spec_index_fullname(spiAcronym)
}
\arguments{
\item{spiAcronym}{The acronym of the spectral index.}
}
\value{
Character. The full name of the spectral index corresponding to 
the given acronym.
}
\description{
This function retrieves the full name of a spectral index based on its acronym.
}
\examples{
fullname <- spt_spec_index_fullname("NBR")

}
