% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_management.R
\name{spt_start_gee_task}
\alias{spt_start_gee_task}
\title{Start GEE Task status}
\usage{
spt_start_gee_task(task, taskTable = NULL, taskTablePath)
}
\arguments{
\item{task}{The task object containing information about the task.}

\item{taskTable}{The task table to update (optional). If not provided, it will be read
from the specified \code{taskTablePath}.}

\item{taskTablePath}{The file path to the task table.}
}
\value{
\code{TRUE} if the update is successful, \code{FALSE} otherwise.
}
\description{
This function updates the status of a task in the task table to indicate that the associated
Google Earth Engine (GEE) task has started.
}
\examples{
task <- spt_get_task(taskTable = myTaskTable, taskID = "12345")
spt_start_gee_task(task, taskTable = myTaskTable, taskTablePath = "path/to/taskTable.csv")

}
