% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_generator.R
\name{spt_status_list}
\alias{spt_status_list}
\title{Get List of Status Values}
\usage{
spt_status_list()
}
\value{
A character vector containing the list of status values.
}
\description{
This function returns a character vector containing a list of status values 
for a specific process.
}
