% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_management.R
\name{spt_task_filename}
\alias{spt_task_filename}
\title{Generate Task Filename}
\usage{
spt_task_filename(task, projectAccronym, versionNumber)
}
\arguments{
\item{task}{A task object containing information about the task.}

\item{versionNumber}{A numeric value specifying the version number.}

\item{projectAcronym}{A character string specifying the project acronym.}
}
\value{
A character string representing the generated filename for the task.
}
\description{
This function generates a filename for a given task based on its properties and the 
specified project acronym and version number.
}
\examples{
task <- spt_get_task(taskTable = myTaskTable, taskID = "12345")
spt_task_filename(task, projectAcronym = "ABC", versionNumber = 1.0)

}
