% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_management.R
\name{spt_task_to_dataframe}
\alias{spt_task_to_dataframe}
\title{Convert Task to Data Frame}
\usage{
spt_task_to_dataframe(task)
}
\arguments{
\item{task}{A task object containing information about the task.}
}
\value{
A data frame with two columns: "taskParam" representing the task parameter names,
and "taskValue" representing the corresponding task values.
}
\description{
This function converts a task object into a data frame, where each column represents a
task parameter and the corresponding value.
}
\examples{
task <- spt_get_task(taskTable = myTaskTable, taskID = "12345")
spt_task_to_dataframe(task)

}
