% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_management.R
\name{spt_update_gee_task}
\alias{spt_update_gee_task}
\title{Update GEE Task}
\usage{
spt_update_gee_task(geeTask, task, taskTable = NULL, taskTablePath)
}
\arguments{
\item{geeTask}{The GEE task object to update.}

\item{task}{The task object containing information about the task.}

\item{taskTable}{The task table to update (optional). If not provided, it will be 
read from the specified \code{taskTablePath}.}

\item{taskTablePath}{The file path to the task table.}
}
\value{
\code{TRUE} if the update is successful, \code{FALSE} otherwise.
}
\description{
This function updates the status and information of a Google Earth Engine (GEE) task in 
the task table.
}
\examples{
task <- spt_get_task(taskTable = myTaskTable, taskID = "12345")
geeTask <- spt_create_gee_task(task)
spt_update_gee_task(geeTask, task, taskTable = myTaskTable, taskTablePath = "path/to/taskTable.csv")

}
