% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_management.R
\name{spt_update_post_task}
\alias{spt_update_post_task}
\title{Update Post-Processing Task status}
\usage{
spt_update_post_task(task, state, taskTable = NULL, taskTablePath)
}
\arguments{
\item{task}{The task object containing information about the task.}

\item{state}{The new state of the post-processing task.}

\item{taskTable}{The task table to update (optional). If not provided, it will
be read from the specified \code{taskTablePath}.}

\item{taskTablePath}{The file path to the task table.}
}
\value{
\code{TRUE} if the update is successful, \code{FALSE} otherwise.
}
\description{
This function updates the status of a post-processing task in the task table.
}
\examples{
task <- spt_get_task(taskTable = myTaskTable, taskID = "12345")
spt_update_post_task(task, state = "COMPLETED", taskTable = myTaskTable,
taskTablePath = "path/to/taskTable.csv")

}
