% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_generator.R
\name{spt_write_tasks_table}
\alias{spt_write_tasks_table}
\title{Write Tasks Table}
\usage{
spt_write_tasks_table(taskTable, taskTablePath)
}
\arguments{
\item{taskTable}{The task table to be written (data frame).}

\item{taskTablePath}{The file path where the task table should be written.}
}
\value{
TRUE if the writing is successful, FALSE otherwise.
}
\description{
This function writes a task table to a specified file path.
Uses a file lock over the to ensure concurrence management.
}
