% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{spt_ymd_date}
\alias{spt_ymd_date}
\title{Get the Current Date in YYYYMMDD Format}
\usage{
spt_ymd_date()
}
\value{
A character string representing the current date in YYYYMMDD format.
}
\description{
This function returns the current date in the format "YYYYMMDD".
}
\examples{
spt_ymd_date()
}
