% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{todaySubDate}
\alias{todaySubDate}
\title{Calculate the date that is a specified number of days before today}
\usage{
todaySubDate(dd)
}
\arguments{
\item{dd}{The number of days to subtract from the current date.}
}
\value{
A date object representing the calculated date.
}
\description{
This function calculates the date that is a specified number of days before the current date.
}
